Make verifier happy

+++ tools/qt/usb_boot_maker/src/main/mainwindow.cpp
@@ -726,7 +726,7 @@ void MainWindow::downloadReplyFinished()
     {
         ++mCurrentApplication;
 
-        if (mCurrentApplication >= applications.length())
+        if ((i64)mCurrentApplication >= applications.length())
         {
             switchToState(UsbBootMakerState::BURNING);
         }
